requirements.txt
colorama==0.4.4
ConfigArgParse==1.5.3
PyYAML==6.0.1
uiautomator2~=2.16.19
urllib3==1.26.18
emoji==1.6.1
langdetect==1.0.9
atomicwrites==1.4.0
spintax==1.0.4
requests~=2.31.0
packaging~=20.9
boto3==1.35.90
python-dotenv==1.0.1